+++ package.json
@@ -7,7 +7,7 @@
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "server-up": "node ./apiRest.js"
+    "server-up": "nodemon ./backend/index.js"
   },
   "private": true,
   "dependencies": {
@@ -32,6 +32,7 @@
     "jsonwebtoken": "^9.0.0",
     "moment": "^2.29.4",
     "momentjs": "^2.0.0",
+    "morgan": "^1.10.0",
     "multer": "^1.4.5-lts.1",
     "mysql": "^2.18.1",
     "rxjs": "~7.8.0",
@@ -53,6 +54,7 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
+    "nodemon": "^2.0.22",
     "typescript": "~4.9.4"
   }
 }
